Make the parser aware of the query type.
It's helpful to know whether we're being called from a Service, Bucket or
Object-mode handler when parsing the request. If we, for example, call
ObjectStatus when we're at the root (Service mode) or at the bucket level
(Bucket mode), we can reject the command before even trying to look for an
object name in the query.