Support presigned URLs.
Turns out we can do this is a minimally invasive way. When we receive the
request, check if the Authorization: header is present. If so, use it. If not,
check if we have instead the correct parameters for a presigned URL. If so,
synthesize an Authorization: header to sent to the Authenticator from the
query parameters.

This is complicated slightly by the presence of two forms of presigned URL.
The first corresponds to AWS v2 signatures and doesn't include the region. For
this form we synthesize a v2 Authorization header.

The second corresponds to AWS v4 signatures and is activated (either from
`s3cmd signurl` or `aws s3 presign`) if a region other than us-west-1 is
configured. This has a region, and gets all the proper fields. Of note are the
elongated Credential= field, and the mandatory SignedHeaders= field.

No signature-specific unit tests yet. This will be fixed, but it's actually
less of a showstopper than one would think, because we ultimately end up
calling the header-based paths anyway.