Re-check the presigned URL expiry time.
Turns out the RGW checks the presigned URL expiry time before it even calls
out to handoff, at least in v17.2.6. Don't trust this; check it ourselves.

There is a toggle to turn this check off if it's seen as a problem. It might;
for v4 there's the conversion via strptime() then mktime(), which isn't
terrible but is unnecessary if rgw has already checked.

Minor other tweaks.