Tests and infrastructure.
Add some substance to the unit test. This is still not where I want it, but
after a conversation with Marco I'm going to merge this with the Handoff
branch and test it for real.

Add a bunch of infrastucture support for buf. Make it so one just has to show
the files involved in the gRPC/protobuf generation, and it automatically
generates variables that allow a custom command to run buf only when
necessary, and to have the generated object libraries rebuilt on demand.