Prep: Rename config to include 'http'.
Since there might be more than one URI, make it explicit to what we're
pointing.

Inferring the mode from the URI itself might seem easier, but I
prefer not to have a clear bool that says which mode we intend. Then we
can have sensible defaults and not have magical-seeming behaviour.