Checkpoint: Compiles and runs with PIMPL'd Helper
Useful milestone: We build and link, whilst limiting the impact to one
file.

This is interesting - it might mean we can actually statically link that
single file against the grpc++ libraries, not the entire library...
Something to consider.

Next step: Running client. This is the important one, as any runtime
gotchas (uncaught ODR violations for example) will show up then.