Checkpoint: gRPC unit tests. Docs.
Add some e2e (ish) unit tests for the gRPC arm of HandoffHelperImpl. This was
quite a bit easier and more convincing than with HTTP, as we didn't need to
inject our own test callout function, we could just spin up a gRPC server on a
random port and call it.

The gRPC test server is still a very simple implementation, but it does
actually authenticate the request, just as for HTTP (in fact it's using the
same code).

Had to change the authentication implementation in the unit test to not take a
base64 stringToSign any more, and have the HTTP server mock decode the JSON
version which _is_ base64 encoded.

Add an implementation of the Status() RPC to the AuthServiceClient for
completeness. It's useful for testing if nothing else.

Add a bunch of missing doxygen.