The Big EAK rename. Always try to send authz.
No more EAK - now we call then AuthorizationParameters (like the protobuf) and
we always send them if we have them, regardless of whether there's a token
HTTP header - we're likely to want them for authorization anyway.

In fact, we're likely to want more for authorization - for example, sending
the full list of query parameters will probably be useful - that's how the
prefix is set for the ls command, and probably for other things too.