Checkpoint: Implement logging prefixes.
Implement DoutPrefixPipe so we can add prefixes to log messages. Implement a
special case where we create a standard prefix for everything we do in the
HandoffHelper - this should make it easier to track down issues in the future.

There's a note to improve the AuthServiceClient implementation - the API
doesn't make it possible to tell the difference between an authentication
problem and a gRPC connection problem.