Fix subtle test problem with gRPC input.
The sigfail_tests group arbitrarily changed some inputs to check that the
authentication failure was propagated. This is fine, as far as it goes.

Thing is, the stringToSign is base64-encoded. So corrupting the input can
produce a stringToSign that is not valid utf-8. gRPC doesn't like that - if
it's not valid utf-8, it doesn't belong in a string, it goes into a 'bytes'
type.

We need to check for genuinely bogus input and catch the exception. However,
we can fix the test by un-encoding the stringToSign, changing the unencoded
version, then re-base64 the changed string. This works fine.