Configuration tweaks.
Make signature v2 support runtime alterable. Support via the observer rather
than by querying each time.

Set up a few more non-runtime alterable variables (rgw_handoff_enable_grpc,
rgw_handoff_enable_presigned_expiry_check) in HandoffHelperImpl::init()
instead of using a runtime lookup. I don't think the performance differential
matters, but it's more sensible to maybe sure that people can't change, say,
the gRPC mode at runtime. That's totally possible, but it's complicated and
pointless.

Note that config changes via the observer don't work in the unit test harness.
I don't know why, and I don't really care - the way I've implemented the
observer pattern means I have to have API in HandoffHelperImpl to change
things in helper anyway; so I can just use that API and rely on the observer
working in the main daemon because it clearly does work.

The HTTP path is not doing anything fancy with the observer and just looks
things up live. This is  because (a) it's too slow for the config lookup to
matter, and (b) it's not long for this world.