Fix the ninja break.
It seems that CMake (version 3.16.3 anyway) generates bogus build.ninja code
for target rgw_grpc_genobj and rgw_grpc_test_genobj. It pulls out a dependency
on an test for an unrelated include file (valgrind/helgrind.h) and adds a
build rule for it for each of those targets. That's wrong; and ninja really
doesn't like duplicate build rules.

Work around this in a rather agricultural way: Build all the grpc stuff into a
single library object. Then the bogus include is still there, but it's only
there once which is fine.

Will try it on a newer CMake at some point.