Make authz param optional. Config lock. Obs test.
Add parameters to control when we send authorization parameters. By default
it's always on. Users can now set 'with token' mode, and 'off' as well.

Add a shared_lock around the runtime-configurable parameters. It'd be a shame
if they changed mid-function.

Add tests for the HandoffConfigObserver utility class. This is a compile-time
polymorphic helper class for HandoffHelperImpl, so we can test it by mocking
the methods of HandoffHelperImpl that we actually use, and there's no runtime
cost.

Test that each of the specifically-documented runtime configurable parameters
respond properly to changes.