Remove bogus include.
As the comment explains, this doesn't help. If we're building fmt ourselves,
target_link_libraries() is the right way to add it. If we're not, this is
pointing at includes that don't necessarily match the system includes, which
is bad. In particular, it seems to confuse the linker at certain times,
possibly because the in-tree version is compiled with a different C++
standard.