Set the right C++ standard for gRPC compile.
We need the C++ library and its dependents to be compiled C++ 17 like the rest
of the source tree. Otherwise it looks like it's worked, but we get subtle
link errors with, say, absl::string and absl::time.

In retrospect it would have been better to have the FetchContent in
src/CMakeLists.txt instead of the toplevel. Oh well.