Use abseil time parsing instead of strptime().
Abseil's date/time parsing is really nice. It's concise and well-defined, and
supports useful things like knowing that 'Z' means the same thing as '+00:00'
in a time string. So it can parse S3 date strings like '20010203T040506Z'
without breaking sweat.

This is still in the (disabled by default) Presigned URL expiry check path,
but it's still a test that fails on a machine in the 'wrong' timezone.