Protobuf changes to explicitly allow anon.
Add a user type to auth.proto, so we can be explicit about anonymous vs.
regular users. The default is a regular user, so the protocol doesn't really
change.

This is likely to be the only survivor of the prototype anonymous authz-only
test implementation. Everything else is going to be redone in a way that
doesn't require as much hackery to get anonymous requests sent to the right
place.

In rgw, 'anonymous' is a special local user. We have no obvious way of
specifying that user as the result of an auth request to a remote user engine
like Handoff. Rather than muck about with the remote user interface, which
will almost certainly lead to future merge nightmares, I'm going to go back
and modify or add a local engine that does what we need without cheating.