Hack anonymous authentication into Handoff via S3.
This branch is here for reference. We're not going to do it this way.

Normally this wouldn't work. The AnonymousEngine would see the request
first, and would accept it; no further authentication attempt would be
made.

The change in rgw::auth::AnonymousEngine::authenticate() (commented out
here) causes that engine to deny the request, so that it goes through
other authentication engines.

The rest is some fairly obvious hackery to make Handoff deal with a
request that clearly doesn't have the right stuff to be an authenticated
S3 request. I don't like the number of possible paths through auth()
that this generates. It's asking for trouble.

More importantly, the return value from an external (remote) auth engine
is different to that from a local engine. The local engine points to a
local user, and anonymous *is* a local user. It might be possible to
have a remote engine return it, but I didn't see it, and I'm not sure
I'd trust it.

In any case, it makes more sense to augment the anonymous engine
(there's an S3-specific anonymous engine which we'll probably modify) or
to add another engine, which by default returns the right type of
result.

Handoff should not require too much pain to be sufficiently reusable for
our purposes here. A separate engine should give us better hooks for
testing too, without needing an explosion of test scenarios to encompass
anonymous mode.