The Big Protobuf Change. WIP. Some tests failing.
- Rename all the things and deal with differences in the messages.

  - Proto is called something else.
  - There's no separate authorization parameters submessage (that will
probably bit us in the arse at some point).
  - The access key is not a separate field, so servers have to extract
it from the authorization header themselves.

- Only send x-amz- headers now.

- Remove the Status() (ping) method.

- Rework the client Auth() method to handle Richer errors, and remove
parse_auth_response() entirely.

We still need to deal with the HTTP status code coming back in the
HandoffAuthResult.

Four tests fail. This appears to be related to not setting the request
method in the tests, which means these tests may well have been failing
before this work, which is dumb.