Fix unit test failures.
The HTTP failures were down to applying gRPC changes to the HTTP code
where they're not relevant. However, it also found a subtle problem with
the sigfail_tests, where I deliberately corrupted the access_key field
but left the correct access key in the authorization header. This meant
that an authentication that used the separated access_key field would
fail, but an authentication that parsed it from the auth header (gRPC
mode) would pass. That's super-confusing - corrupt both access keys.

The gRPC failures were down to how the failure of the gRPC is parsed.
This is good testing - we asserted a difference between a transport
failure and an authentication failure, and that got broken in the
transition to endpoint gRPC, and the tests broke as a result. Yay.

Now, if we get a failed authentication RPC we look for an S3ErrorDetails
message embedded in the grpc::Status error details. If we don't find it
for any reason, we return a TRANSPORT_ERROR enum, on the assumption that
our errors always have an S3ErrorDetails unless the service was
unreachable or broken. The tests look for this.