Return a good error code to RGW.
What RGW does with it from there isn't up to us, alas.

Translate the detailed error code returned by the Authenticator into a
matching RGW error code, and return that to RGW. If we get an
Authenticator code we can't match to an RGW code, return something
appropriate that should give the HTTP code that the Authenticator
specifies.

It seems RGW doesn't respect the code we give it, but we're doing our
best to give as much good information as possible.