The real problem - a missing HAVE_ABSEIL.
opentelemetry-cpp uses target_compile_definitions() to propagate
configuration variables. In our case we care about HAVE_ABSEIL - if
that's present we'll use regular abseil headers, otherwise we'll use an
in-tree version. These will clash if there are abseil headers elsewhere,
which there are when we build with gRPC.

We need to build otel with -DWITH_ABSEIL=ON so we don't have multiple
versions of the same headers. For otel itself this works fine, but we
also build src/common/tracer.cc using opentelemetry, and that needs the
definition too, otherwise it tries to use those same clashing headers.