OBJGEN1-627: In-tree libxml, lower thresholds.
- Pull in libexpat using FetchContents, instead of using find_package()
for the system version.

- Work around CMake < 3.18 lack of SOURCE_SUBDIR. This essentially means
we have an add_subdirectory() that we wouldn't need with newer CMake.
(SOURCE_SUBDIR is used when the CMakeLists.txt file isn't in the root of
the source tree, as is the case for libexpat.)

- Call the Billion Laughs attack protection APIs to lower the mitigation
thresholds /very/ aggressively. We really should not be getting
complicated XML and I would argue we could lower the settings further.

- Fail the build if libexpat is not >= 2.5.0. This is a guard against
future error.