Add a unit test for in-tree expat. WIP.
Make sure we can compile and link against an appropriate version of
libexpat. This test can stay in even if we end up on a system that has
up-to-date libexpat installed, because we're just testing that the
Billion Laughs mitigations do what we expect.

More to come. At the moment it just makes sure we can compile and link.