Build gRPC and OTel with standalone abseil-cpp
Commit to having a standalone abseil-cpp installation. This is the right
way; using the abseil baked into gRPC feels really brittle to me.

So: gRPC is straightforward, just add the necessary prefix path.

OTel is modified so if we're in gRPC-capable mode, set the various flags
and prefixes required to turn off the in-tree abseil.