Port gRPC integration only from v17.2.6 branch.
Port just the gRPC integration from branch
atl-v17.2.6-add-handoff-auth-integrate-grpc, via xxx-prebuilt-cmake (the
latter is misnamed, it should be 'xxx-prebuilt-grpc'.

- Add switch WITH_RADOSGW_GRPC CMake toggle. Default to ON.

- Add toplevel CMake to pull in prebuilt abseil-cpp and gRPC.

- Add CMake library code (cmake/modules/RGWgRPC.cmake) to make automatic
build of gRPC and protobuf sources, as generated by the buf tool.

- Add buf generation .yaml files and the .proto definitions for testing.
Add the authentication protobuf here too, it does no harm as it's not
used yet.

- Add unit test src/test/rgw/test_rgw_grpc.cc and a utility library
include file. This is a link and execute test for the gRPC integration.

The gRPC integration is documented in detail here:
  https://kb.linode.com/display/Development/gRPC+integration+for+RGW

This document needs updated to reflect the prebuilt gRPC and abseil,
which are new and are present because the previous in-tree build, whilst
effective, was slow and was incompatible with OpenTelemetry which we
need to have for gen2.