Port Handoff auth from the v17.2.6 branch.
Port the remainder of Handoff from branch
atl-v17.2.6-add-handoff-auth-integrate-grpc, via xxx-prebuilt-cmake (the
latter is misnamed, it should be 'xxx-prebuilt-grpc').

- Add all the handoff configuration options, as documented. Note that
several are tagged as 'runtime'.

- Add handoff sources (rgw/rgw_handoff.cc, rgw/rgw_handoff_impl.cc) and
their corresponding headers, and the handoff unit tests
(test/rgw/test_rgw_auth_handoff.cc) which uses the utility library
installed with with gRPC change.

- (rgw_auth.cc) Allow automatic user creation to be disabled if desired.

- (rgw_auth_s3.h) Add class members for the handoff 'engine', analogous
to the LDAP engine.

- (rgw_common.cc, rgw_common.h, rgw_log.cc, rgw_user.cc) Add display
types for handoff, again just like LDAP.  There's a chance future
updates will clash with these - we'll cross that bridge if we come to
it.

- (rgw_main.cc) Like LDAP, cleanly shut down the engine at daemon exit.

- (rgw_rest_s3.cc, rgw_rest_s3.h) The main integration point.

  - Handoff needs to be included in the list of viable authentication
  methods without which the daemon will exit.

  - class HandoffEngine, the interface through which rgw calls out to
  Handoff. This class is described in the documentation.

  - One small reference in the STS engine so that where it references
  the LDAP engine, it now also references Handoff. We don't expect this
  to be used.

- (rgw_token.h) A separate authentication token for Handoff, analogous
to LDAP.

- (store/dbstore/CMakeLists.txt) A small but important bug fix, where
dbstore incorrectly pulls in the wrong include file for libfmt. This
should probably be upstreamed, there's no case where this is the correct
thing to do.

- (test/rgw/CMakeLists.txt) Add the Handoff unit test.

Handoff is documented in detail here:

  https://kb.linode.com/display/Development/RGW+%27Handoff%27+authenticator+plugin