Add error code to Result type. More tests, docs.
Add an integer code to the UBNSClient::Result type. This code is all we can
return to RGW to influence the error message the user sees, so we'd better
have a way of communicating it.

Make the success-or-failure construction of the Result type more obvious. It's
a little more verbose, but it's totally clear what's being done. The
constructors are now private, and honestly I think it's a better API.

Switch the default URI to localhost:9000. It was a pasto before.

Bring in the config observer tests from Handoff. They work well.

Add some more doxygen comments.