Wire StoreQuery into REST mgrs and handlers.
Wire a flag to enable or disable StoreQuery into RGWRESTMgr_S3 so we
don't have to hit the config proxy for every operation. This takes the
form of a construction-time flag and a member variable, and can be
queried quickly at op processing time.

A lot of the diff is in adding a bool to a bunch of signatures to pass
the storequery enable status to where it's needed.

When SQ is enabled, for handler types Service, Bucket and Obj, create an
sq handler and use it to detect if the sq HTTP header is present. If so,
either return a storequery RGWOp subclass to handle the storequery
request, or return nullptr to indicate that op detection should
continue without further sq considerations.

If the op parsing throws an exception, that gets mapped onto a nullptr
return from RGWHandler_REST_*_S3, indicating a failure to find an
appropriate handler. The user will see an error, typically 405 method
not allowed.

Note that most operations get very simple processing - if the storequery
header isn't present, return nullptr and carry on with normal
processing.

The op processing is admittedly complex, but it's arranged so that we
don't need separate storequery RGWOp subclasses for each type of
operation. It currently doesn't matter, but if we add items later on
that need to process multiple op types (get, head, post etc.) we can be
parsimonious with our RGWOp types.

Likewise, the use of the single RGWHandler_REST_StoreQuery_S3 class
(with an op type in the constructor) instead of a handler for each type
(service, bucket, object) allows us to have a single class that handles
all the request types. It's fair criticism that it doesn't look like any
of the other handlers, but the other handlers aren't likely to need to
handle multiple request types in quite the same way.