Significant integration changes to storequery.
Remove the multiple op_get() patches, and instead patch
RGWRESTMgr_S3::get_handler() to return our handler type in the proper
place. This is a lot easier to understand in the overall flow.

We still create the handler with the query type. This feels a little
vestigial, but it works out nicely as it allows us to save some work
later on when deciding whether or not to load buckets and objects. We
don't always have to do that.

No longer throw in the handler's op_get() implementation. The exception
is specific to the old interface, and has no place in the regular
interface. It means all error returns are nullptr, and you'll get
'method not allowed'. The logs will tell the real story.

Implement RGWHandler::init_permissions(). This is very important;
init_permissions(), despite its name, loads a bucket and optionally an
object.  We can't call a superclass's version because we don't want all
those nasty permissions checks, so we've reimplemented a subset here.

Add a fair bit of doxygen.

Smaller changes:
- In tokenize(), wrap boost::tokenizer calls in try/catch. It can throw.
- In …ObjectStatus::execute() assert that the bucket is set.