Checkpoint: Create and Delete state machine fw.
Start building out RAII-safe state machines for create and delete. Since
RGWCreateBucket::execute() and RGWDeleteBucket::execute() have multiple
error returns, to keep the code maintainable we need our rollback
actions in the destructor, so that we're guaranteed to be called.

This in turn means the destructor needs to know where in the flow we
are.

This in turn means a state machine. So implement that. Make it a
template so we can mock the gRPC implementation class out and test it
somewhat standalone.

Pull the Result class out of UBNSClient, there's no reason for an inner
class and it makes templating the state machines much harder.

Implement skeletons of state object interactions in
RGWCreateBucket::execute() and RGWDeleteBucket::execute().