Checkpoint: Testable state machine.
Template-ise the state machine so it can be easily tested. Create a mock in
the unit test and check the machines' transitions.

More tests to come. At the moment, the obvious things are working again, but
with the advantage that there's an automatic rollback if some operations fail
mid-create or mid-delete.