Assert if the user attempts to use an invalid state.
There are some states the user can select, and some that only the state
machine can select. If the user requests a machine-only state, call
ceph_assert with a helpful message.

Test this, but disable the test by default (without the use of an
environment variable) because Ceph 17's gtest is too old to have
threadsafe death tests.