Refactor channel setup.
Don't make it up to the caller to determine whether or not the channel
is mTLS-enabled. Set that once (in init) and have new method
set_channel() look it up when necessary, and call renamed
_set_insecure_channel() and _set_mtls_channel() do the work.

This makes it much less likely that we'll accidentally switch into
insecure mode somewhere in the daemon's lifetime.