Update .proto to add owner. Create bug. Logging.
Due to problems found in QA, the upstream .proto has been updated to require
an owner field for Update and Delete requests as well as Add requests.

In addition, the requirement has been tightened to include the requirement
that the owner field needs to be the owner of the bucket as seen by Ceph. For
the Create machine this is straightforward as the owner is the user making the
request. For the Delete machine, we have to fetch the owner from the bucket.
This is easy too (it's already loaded in to the req_state) but it is a new
feature.

Add explicit logging of the (bucket,clusterid,owner) triple almost everywhere.
This should make it easier to debug problems that might come from differences
in any of those fields. Most of those entries are at log level 5.

Fix a bug in the Create state machine that was using the owner field instead
of the cluster ID.

Now Add, Update and Delete all take the cluster and the owner, and all the
APIs and tests have been updated to reflect this.