Set skip_timestamp_validation for presigned URL requests.
As requested by the Authenticator team, set the field
'skip_timestamp_validation' in AuthenticateRESTRequest if we've
synthesized an Authorization header from a presigned URL request.
Without this, v2 presigned URL requests are not formatted exactly
correctly and fail to authenticate.

This isn't a security problem! RGW validates the expiry time of the
presigned URL before ever calling handoff. This just allows us to
externally authenticate the request without a spurious failure.