Updated error responses. Idempotent Create.
Support the updated error response table from the devs. I'll update the
docs once I get confirmation that this is the final table.

Support idempotent create. Essentially, if the backend says this is an
identical create of a bucket, let it progress. However, don't do the
Update processing, because that won't work.

This has required the addition of a new Create machine state
CREATE_RPC_SOFT_FAILURE. In essence, we have a state that allows the
RGWCreateBucket::execute() function to continue to the point where it
discovers the bucket already exists, then the destructor knows that we
can safely move from CREATE_RPC_SOFT_FAILURE -> COMPLETE and log
success.

Documentation updates will be required. I'll get to it.