Update mock UBNS server for idempotent create.
Previously it wasn't necessary to store much state in the mock UBNS
server in the unit tests. We just stored a simple state enum for the
bucket.

Now, we want to check that the return codes for an idempotent create are
handled, so we need to be able to tell the difference between a create
with the same owner and cluster (which is allowed) and a create with a
different owner or cluster (which is not).

So, we have to store the owner and cluster as well as the bucket state.
This changes the API which touches a lot of code in the test. However,
the bulk of the change is quite minor: When deciding whether or not we
should return the 'idempotent create, continue' message, check if the
cluster ID and owner agree.