Change create behaviour on SOFT_FAILURE.
Previously, if we got to state CREATE_RPC_SOFT_FAILURE, we'd allow
RGWCreateBucket::execute() to proceed but wouldn't sent the update.

It appears that's unwanted behaviour - my take was that we wanted an
idempotent create which means we could safely allow the RGW create code
to continue. However, we *don't* want the bucket created on a different
cluster so we need to exit RGWCreateBucket::execute() immediately and
return a 200.