Add UBNS callout for Admin REST API delete_bucket.
Turns out some Linode integrations use the admin REST API for some
things. As far as I'm aware (I asked Vicki) the only thing that matters
for UBNS is remove_bucket. That's implemented in
RGWOp_Bucket_Remove::execute(), in src/rgw/rgw_rest_bucket.cc.

It's not an accident that the new code is near-identical to
RGWDeleteBucket::execute(). There's no real reason to change it.
However, if this gets used and as testing proceeds it /might/ be
necessary to tweak op_ret.