Fix bad interaction of StoreQuery and anon authz.
v1 StoreQuery (which will be replaced with gRPC after the gen2 beta)
uses bare HTTP requests to the REST server. This works fine, as far as
it goes. However, they are by definition also anonymous requests, which
means that they're candidates for anonymous authorization if anon authz
is enabled.

That's wrong. The storequery requests probably won't authorize, but
that's not the point! The purpose of storequery is to bypass the
authentication and authorization to do a quick 'object present' check
for the microservice platform.

So: In addition to existing checks, look to see if this looks like an
HTTP StoreQuery request (=> it has the x-rgw-storequery: HTTP header)
and if so, skip anon authz.

Add a more detailed comment to explain the three-part if block around
anon authz. It's getting cluttered.