Authz infra start. Add z state to req_state.
- Add HandoffAuthzState class, add unique_ptr to req_state and init in
  rgw_process.cc.

  Take the opportunity to set up rgw_handoff_fwd.h for forward
declarations and remove crufty fwd in rgw_common.h.

  Note the include in rgw_common.cc so the constructor for req_state
knows the size of HandoffAuthzState (unique_ptr needs the deleter, and
the deleter needs the size). This is similar to the PIMPL massaging for
HandoffHelperImpl.

- When rgw_process.cc initialises the HandoffAuthzState, pass a
  possibly-null shared_ptr to the HandoffHelper. This keeps all the
Handoff-related initialisation code in the handoff source files, rather
than having to put Handoff-specific calls in rgw_process.cc.

- Remove 'local_authorization_bypass_allowed()' associated with old,
  dead config option.