Refactor gRPC channel support.
We'll shortly be needing two gRPC channels, one for authn and one for
authz. Previously, the gRPC channel information was represented as
fields (including a lock) on HandoffHelperImpl. Refactor the channel
manipulation into class HandoffGRPCChannel with an integral lock, and
manipulate via member object authn_channel_ (via get_authn_channel()).

Already added corresponging authz_channel_ / get_authz_channel().

This simplifies things quite a lot. I imagine it will especially
help when the mTLS stuff has to be brought over from UBNS, but the less
said about that today the better.

This required some changes to the config observer interface. Recall that
we have a template that allows us to test runtime configuration changes
without having to mock up the entire helper. This was based on the old
direct access interface and needed changed to match the sub-object. It
simplified that interface too, though we needed to mock
HandoffGRPCChannel (trivially) as well.