Add stubs in HandoffHelper* for verify_permission.
Add non-functional endpoint verify_permission() in HandoffHelper
(calling through to HandoffHelperImpl) that will do the RPC and call the
authorizer.

As we're calling via the helper, use the RGW values
(rgw::IAM::GetObject etc.) instead of anything remotely gRPC-involved.
We'll do the translation in HandoffHelperImpl.

The current stub returns 0 (success). This allows me to try control flow
without having to worry about wiring in the gRPC.