Start unit testing the gRPC parts.
Mock up a gRPC server inside the unit tests. Boy, am I glad I abstracted
the GRPC test server code a while back.

The mock actually starts an authn and authz gRPC server. This is easily
done and not intrusive, so it's not worth changing the
HandoffHelperImpl::init() signature just to accomodate a partial channel
startup.

Authorize() is stubbed out, but Ping() is actually doing all it has to -
it copies the request authorization_id() into the response and returns
ok. Again, I like having a Ping() RPC because it means we can test stuff
without all the super fiddly marshalling that Authorize() is bound to
require.

But that's next up!