A great deal of tedious gRPC groundwork and tests.
This is a lot of the work we will need to start sending and receiving
RPCs.

- Add all the fields we need to HandoffAuthzState. This will be used to
  populate the AuthorizeRequest later.

- Note that we have fields with the bucket name and object key name in
  HandoffAuthzState. This is to allow us to unit test a lot of this code
without having to mock rgw::sal::Bucket and rgw::sal::Object, which is a
lot more trouble than it would be worth. However, we MUST remember to
populate those fields in the handoff_authz->enabled() paths, once any
bucket and object has been loaded.

- It occurs to me that we'll probably have to have fields for source
  object as well at some point. Cross that bridge...

- Make the protobuf messages and the AuthorizerClient::AuthorizeResult
  field ostream printable. After a lot of very tedious manual
implementation, I switched to protobuf's JSON printer which is very good
and almost certainly a great deal faster than ostream operator<<.

- Add maps in both directions from rgw::IAM opcodes to protobuf S3Opcode
  enum. I know there are shortcuts, but I'm not confident they'll always
be correct so I'm playing it safe.

- Some little unit tests to check the marshalling of stuff into
  AuthorizeRequest, which is very fiddly. Add PopulateAuthorizeRequest
which we'll certainly use in the op implementation path.

- The AuthorizeBasics test is currently failing. I've not implemented
  the mock gRPC server Authorize() processing yet, today has all been
writing groundwork for that and the real client implementations.