Pass extra authentication fields to authz_state.
When authenticating, set all four of the string fields returned by the
Authenticator, for reflection to the Authorizer later.

This is complicated by the fact that the internal APIs call the
authentication functions with a const req_state*. We need to set fields
in req_state.authz_state.

Deal with this by putting the fields we need to modify in a struct, and
marking the struct as 'mutable'. This is a hack; I admit it. However,
I'm not modifying the authenticate() API just for this. For the most
part I completely agree with the req_state* being const. Here, we need
to save some information, but we're doing it 100% with safe C++
functions (std::string). I think the compromise is acceptable.

Add unit test support for checking we're setting these values.