WIP: Switch to richer error model.
Ok, we're going to the Richer Error Model, to keep in style with other
components. This is quite painful, actually.

So: A non-error means 'Allow'. Anything else means access denied, more
data required, or some other error.

In the process I've moved the ExtraDataSpecification from the response
to the AuthorizationErrorDetails message. This sort of makes sense,
because it's not necessary if access is allowed, and if not there's an
error message which can contain it.

In the process I've broken the unit tests that call
PopulateAuthorizeRequest, because they are now trying to load the
attributes on buckets and objects. This was in a worthy attempt to unify
the Populate code with the update in the hhi::verify_permission(), but I
need to introduce a way to mock the data loads to make this work.

Stopping for today. Brain fried.