Set bucket and object names in authz state.
If available, set the bucket name and object key name in the authz state
explicitly, before processing op->verify_permission().

This is really for unit tests, again to obviate the need to mock the SAL APIs.
If we know we can use the authz state names, we can set those easily in unit
tests, and our code can use those copies so we're testing the appropriate
paths.

Note the double check for bucket, then bucket_name not being empty. The first
time I tried to use this without the second check (for RGWCreateBucket) it
failed, because the bucket object was empty, but s->bucketname was set! This
is why we have to visit every RGWop subclass; nothing is ever simple.