Handoff::RGWPutObj::verify_permission().
Implement handoff-aware authorization for RGWPutObj.

Simplified enormously by the knowledge (thanks Ety) that in gen2 we don't
have to deal with a copy_source, in fact if we get that and handoff's enabled
we should fail with an error. We now do that.

This is one of those verify_permission() methods that does things that we will
need later on as well as permissions work. So we needed to include those in
our handoff-only stanza too. This is a pain for the future, but the
alternative (and it's not a good alternative) is a long line of if statements
blocking stuff out, which would have been hard to follow.