HandoffAuthzState requirements stack.
The (minimal) unit test for this was submitted earlier by mistake.

Have a simple stack for authz requirements. This is intended for the case
where we have an op that makes multiple (not extra data-related)
verify_permission() requests to the Authorizer. It would be easy to get
confused if those requests have differing extra data requirements, so make it
easy to just save the existing set of ed requirements and get them back later
on.

The intended use case for this (RGWPutObj) actually turned out to not need it,
because we don't support source object puts (a.k.a. copy-from-source-object)
in gen2. I've left the stack in place just in case it's useful for some other
operation.