Error and log tweaks. More tolerant obj tag load.
Make logging a bit easier to read, especially around richer error model return
values. If we get a richer error model return, stringify the code enum. That
turns out to make a substantial difference to readability.

Don't be so quick to return fatal errors when reading object tags. At the
moment we're getting errors in _load_objtags_from_sal() more than we
organically would because the Python server is dumb and is requesting extra
data when it shouldn't. However, we can't fix that right now, so we'll just
accept ENOENT when trying to read the object tags, return an empty list of
tags, and move on. The worst it can do is submit an empty tag list to the
Authorizer.