Deal with the batch API proto change.
Everywhere we expected a single question and answer, deal with a
list of questions and answers.

Bumped into some odd protobuf bugs, where using repeated fields seems to
cause spurious crashes. That manifests in taking iterators out of
std::optional<message>, it just doesn't work and seems to segfault. The
crappy solution is to copy the value out of the optional, which is just
ugly.