Add multi-value verify_permissions() API.
Now we can ask multiple questions on the wire, we need to give the RGWOp
subclasses the ability to use it. Add
HandoffHelper::verify_permissions(), taking a vector of opcodes and
returning a vector of codes. Point HandoffHelper::verify_permission() at
the new call.

Rationalise the result code handling at the end of verify_permissions().
(Was at the end of verify_permission().) It's a lot easier to read now.
Return an array of RGWOp-compatible result codes if we're not returning
hard errors.