Handoff: RGWDeleteMultiObj.
This is a more complex operation. The basic verify_permission() implementation
is similar (but not identical) to RGWDeleteObj's v_p(), but it only checks the
delete permission on the bucket. (I think - it's not entirely obvious to me).
However, an additional method, handle_individual_objects(), is called on
/every/ object in the request body XML. This does authorization, so we have to
make a handoff call for every object in the request. This is not going to
speed it up!

We also have to test this in rgw_beast_enable_async=true mode, because by
default each object is deleted in a separate boost thread context. We need to
check that that doesn't break the gRPC code.