Target and reqs stacks. µservice filter toggle.
Add a target (bucket and key names) stack to the already-existing requirements
stack. Add push() and pop() that operate on both stacks. They're not going to
be used very often, but it only takes one use.

Add config option rgw_handoff_authz_reject_filtered_commands (defaults to
true) that signifies whether or not RGW attempts to authorize commands that we
expect to be filtered away from RGW by the microservices platform. Whilst it
reduces our work to not have to implement them, it also makes testing much
harder because any testing has to include the microservices platform. This
way, we can set RGW up to attempt to authorize everything.

We still need to add a method to access the filter toggle from the
HandoffHelper (it's currently only in HandoffHelperImpl).