API to add a suffix when forming the authz ID.
The authorization ID (in the question AuthorizationCommon message) we use the
trans_id field of req_state. We supplement it with the question index, and a
'subrequest' index if it's a secondary request in response to an
EXTRA_DATA_REQUIRED response.

That's all fine if we only make a single verify_permission() /
verify_permissions() call per S3 request. Most of the time, that's the case.
When we send more than one (e.g. RGWDeleteMultiObj, which sends an initial
request, then one request per named object in the XML body), we have to have a
different authz ID for each request and using the trans_id field isn't enough.

Rather than lose the link between the authorization ID and the trans_id (which
I think will be useful when diagnosing faults later), I've added the ability
to set a string suffix in the authz state. This can just be a number, but it
can be anything - for example, marking the source and destination objects in a
copy, or numbering the items in a delete-objects request.